Enable phpcs for tests directory
