+++ .phpcs.xml
@@ -26,7 +26,6 @@
   <arg name="severity" value="1" />
 
   <!-- Exclude a few directories and autogenerated files. -->
-  <exclude-pattern>tests/</exclude-pattern>
   <exclude-pattern>build/</exclude-pattern>
   <exclude-pattern>vendor/</exclude-pattern>
