+++ tests/class-test-case.php
@@ -1,4 +1,10 @@
 <?php
+/**
+ * Create WordPress Plugin Tests: Base Test Class
+ *
+ * @package create-wordpress-plugin
+ */
+
 namespace Create_WordPress_Plugin\Tests;
 
 use Mantle\Testkit\Test_Case as TestkitTest_Case;
