--- tests/feature/test-example-test.php
@@ -1,14 +1,0 @@
-<?php
-namespace Create_WordPress_Plugin\Tests\Feature;
-
-use Create_WordPress_Plugin\Tests\Test_Case;
-
-/**
- * Visit {@see https://mantle.alley.co/testing/test-framework.html} to learn more.
- */
-class Example_Test extends Test_Case {
-	public function test_example() {
-		$this->assertTrue( true );
-		$this->assertNotEmpty( home_url() );
-	}
-}
