+++ tests/unit/class-example-unit-test.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * Create WordPress Plugin Tests: Example Unit Test
+ *
+ * @package create-wordpress-plugin
+ */
+
+namespace Create_WordPress_Plugin\Tests\Unit;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * An example unit test suite.
+ *
+ * @link https://mantle.alley.com/testing/test-framework.html
+ */
+class Example_Unit_Test extends TestCase {
+	/**
+	 * An example unit test. In practice, this should be updated to test a function in isolation.
+	 */
+	public function test_that_true_is_true() {
+		$this->assertTrue( true );
+	}
+}
