--- tests/unit/test-example-unit-test.php
@@ -1,13 +1,0 @@
-<?php
-namespace Create_WordPress_Plugin\Tests\Unit;
-
-use PHPUnit\Framework\TestCase;
-
-/**
- * Visit {@see https://phpunit.de/documentation.html} to learn more.
- */
-class Example_Unit_Test extends TestCase {
-	public function test_that_true_is_true() {
-		$this->assertTrue(true);
-	}
-}
